+++ packages/blockera/php/functions.php
@@ -351,3 +351,33 @@ if (! function_exists('blockera_get_filesystem')) {
 		return $wp_filesystem;
 	}
 }
+
+if ( ! function_exists( 'blockera_add_inline_css' ) ) {
+
+	/**
+	 * Adding computed css rules into inline css handle.
+	 *
+	 * @param string $css the provided css from outside.
+	 *
+	 * @return void
+	 */
+	function blockera_add_inline_css( string $css ): void {
+
+		if ( empty( $css ) ) {
+
+			return;
+		}
+
+		add_filter(
+			'blockera/front-page/print-inline-css-styles',
+			function ( string $older_css ) use ( $css ): string {
+
+				if (false !== strpos($older_css, $css)) {
+					return $older_css;
+				}
+
+				return $older_css . $css;
+			}
+		);
+	}
+}