+++ packages/wordpress/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+### New Features
+- Added a new feature to cleanup inline styles from the block elements on the front end.
+
 ## 1.0.1 (2024-12-21)
 
 ### Bug Fixes